Update randomlyAddTaxa
Added new functionality and simplified inputs for users. Also added an example to the .Rd help file. Now takes a single data frame of all taxa, and determines what is missing from the tree, rather than needing this specified apriori. Also prints missing taxa to screen, and automatically converts factors to characters. To do: add warning in case species to be added are from taxonomic groups that are not present in the "real" tree.